pyproject.toml
[project]
name = "FumeWeb"
version = "1.0.0"
description = "Web dashboard for all the FumeStop services."
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "better-ipc",
    "discord-py>=2.4.0",
    "quart>=0.19.9",
    "quart-wtforms>=1.0.3",
    "quartcord>=1.0.1",
]

[tool.uv]
dev-dependencies = [
    "pre-commit>=3.5.0",
    "ruff>=0.8.3",
]

[tool.uv.sources]
better-ipc = { git = "https://github.com/MiroslavRosenov/better-ipc/", rev = "efb145d9dabc0a90e4c6e32be08bd03ae4257249" }

[tool.ruff]
line-length = 85
target-version = "py313"

[tool.ruff.lint.isort]
section-order = ["future", "typing", "standard-library", "third-party", "first-party", "local-folder", "config"]
force-wrap-aliases = true
combine-as-imports = true
length-sort = true

[tool.ruff.lint.isort.sections]
"typing" = ["typing"]
"config" = ["config"]
